// src/app/models/updateVendors.ts
export class updateVendor {
    vendorId: string;
    userId: string;
    vendorName: string;
    gsTNumber: string;
    tanNumber: string;
    emailId: string;
    firstName: string;
    lastName: string;
    phoneNumber: string;
    alternateNumber: string;
    contactPerson: string;
    bankAccountNumber: string;
    ifscCode: string;
    bankBranchDetails: string;
    address1: string;
    address2: string;
    country: string;
    city: string;
    state: string;
    pincode: string;
    tenantId: string;
}